E-commerce Platform Data Analysis Tool/main/requirements.txt
DrissionPage>=4.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
baidu-aip>=4.16.0
openai>=1.0.0
requests>=2.28.0
pandas>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
chardet>=5.0.0